+++ ando/tests/test_Rules.py
@@ -1,5 +1,5 @@
 import unittest
-import AnDO_engine as ando
+import engine as andoE
 class Test_Rules(unittest.TestCase):
 
     def setUp(self):
@@ -8,36 +8,36 @@ class Test_Rules(unittest.TestCase):
     '''testing session rules'''
     def test_session_rules(self):
         list=["sess-180116_001_m_enya_land-001"]
-        self.assertEqual(ando.is_session(list), True)
+        self.assertEqual(andoE.is_session(list), True)
     def test_session_rules_1(self):
         list=["180116_m_enya_land-001"]
-        self.assertEqual(ando.is_session(list), False)
+        self.assertEqual(andoE.is_session(list), False)
     def test_session_rules_2(self):
         list=["180116_001_m_land-001"]
-        self.assertEqual(ando.is_session(list), False)
+        self.assertEqual(andoE.is_session(list), False)
 
 
     '''testing source rules'''
     def test_source_rules(self):
         list=["source"]
-        self.assertEqual(ando.is_source(list), True)
+        self.assertEqual(andoE.is_source(list), True)
     def test_source_rules_1(self):
         list=["Source"]
-        self.assertEqual(ando.is_source(list), False)
+        self.assertEqual(andoE.is_source(list), False)
     def test_source_rules_2(self):
         list=["sources"]
-        self.assertEqual(ando.is_source(list), False)
+        self.assertEqual(andoE.is_source(list), False)
 
     '''testing subject rules'''
     def test_subject_rules(self):
         list=["sub-001"]
-        self.assertEqual(ando.is_subject(list), True)
+        self.assertEqual(andoE.is_subject(list), True)
     def test_subject_rules_1(self):
         list=["sub_001"]
-        self.assertEqual(ando.is_subject(list), False)
+        self.assertEqual(andoE.is_subject(list), False)
     def test_subject_rules_2(self):
         list=["Sub_001"]
-        self.assertEqual(ando.is_subject(list), False)
+        self.assertEqual(andoE.is_subject(list), False)
 
 if __name__ == '__main__':
     unittest.main()
